Uses store for managing JWT
