+++ src/crypt-model.ts
@@ -1,5 +1,6 @@
 import {describeEnum, stringEffectFields, stringFields} from 'faora-kai';
 import {z} from 'zod';
+import {type Result} from './railway.js';
 
 const timeUnit: Record<string, string> = {
   seconds: 'Seconds',
@@ -87,3 +88,4 @@ export type CryptModel = z.infer<typeof schema>;
 export type CrypLizardCypher = z.infer<typeof lizardCypher>;
 export type CryptEncryptionStrength = z.infer<typeof strength>;
 export type CryptIdPayload = z.infer<typeof idPayloadSchema>;
+export type CryptSignResult = Result<string, string>;
